+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ Release Notes
 * Storage: Support for minimum TLS version.
 * Virtual Machine: Support for customData on osProfile Properties 
 * WebApp: Added support for PrivateEndpoints
+* SQL Azure: Support for minimum TLS version.
 
 ## 1.6.1
 * Web App: Workaround ARM regression when Identity is set to "None".