+++ RELEASE_NOTES.md
@@ -1,13 +1,15 @@
 Release Notes
 =============
 
+## 1.6.3
+* Service bus: Add support for authorization rules.
+
 ## 1.6.2
 * Functions: Support Elastic Premium SKUs for Functions service plans.
 * SQL Azure: Support for minimum TLS version.
 * Storage: Support for minimum TLS version.
 * Virtual Machine: Provide control over the public IP
 * Virtual Machine: Support for customData on osProfile Properties 
-* Service bus: Add support for authorization rules.
 * Virtual Network: Add support for vnet peering
 * WebApp: Added support for PrivateEndpoints